requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23

# AI/ML dependencies
langchain==0.0.350
langgraph==0.0.40
openai==1.3.0
google-cloud-aiplatform==0.34.1
google-cloud-speech==2.24.0
moviepy==1.0.3
opencv-python==4.8.1.78
librosa==0.10.1

# Database
sqlite3

# Utility dependencies
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0